requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
aiosqlite==0.19.0
httpx==0.26.0
python-dotenv==1.0.1
pydantic==2.6.1
python-multipart==0.0.9
